requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
requests==2.32.3
google-cloud-documentai==3.5.0
pytest==8.3.5
httpx==0.28.1
